Migrate kubebuilder project to v4
The notable change of v4 is to move controllers into
internal/controller.

Fix #72
