+++ Dockerfile
@@ -1,10 +1,32 @@
+# Build the manager binary
 FROM golang:1.20-bullseye as builder
+ARG TARGETOS
+ARG TARGETARCH
 
-WORKDIR /work
+WORKDIR /workspace
+# Copy the Go Modules manifests
+COPY go.mod go.mod
+COPY go.sum go.sum
+# cache deps before building and copying source so that we don't need to re-download as much
+# and so that source changes don't invalidate our downloaded layer
+RUN go mod download
 
-COPY . .
+# Copy the go source
+COPY cmd cmd
+COPY constants constants
+COPY controller controller
+COPY internal internal
+COPY main.go main.go
+COPY probe probe
+COPY runners runners
+COPY types types
 
-RUN go build -o pie
+# Build
+# the GOARCH has not a default value to allow the binary be built according to the host where the command
+# was called. For example, if we call make docker-build in a local env which has the Apple Silicon M1 SO
+# the docker BUILDPLATFORM arg will be linux/arm64 when for Apple x86 it will be linux/amd64. Therefore,
+# by leaving it empty we can ensure that the container and binary shipped on it will have the same platform.
+RUN CGO_ENABLED=0 GOOS=${TARGETOS:-linux} GOARCH=${TARGETARCH} go build -a -o pie main.go
 
 FROM ubuntu:20.04
 
@@ -12,7 +34,7 @@ RUN apt-get update \
     && apt-get install -y --no-install-recommends fio jq \
     && rm -rf /var/lib/apt/lists/*
 
-COPY --from=builder /work/pie /
+COPY --from=builder /workspace/pie /
 
 EXPOSE 8080/tcp
 EXPOSE 8081/tcp
