+++ PROJECT
@@ -1,6 +1,10 @@
+# Code generated by tool. DO NOT EDIT.
+# This file is used to track the info used to scaffold your project
+# and allow the plugins properly work.
+# More info: https://book.kubebuilder.io/reference/project-config.html
 domain: topolvm.io
 layout:
-- go.kubebuilder.io/v3
+- go.kubebuilder.io/v4
 projectName: pie
 repo: github.com/topolvm/pie
 resources:
@@ -14,10 +18,5 @@ resources:
   group: storage
   kind: StorageClass
   path: k8s.io/api/storage/v1
-  version: v1
-- controller: true
-  group: core
-  kind: Pod
-  path: k8s.io/api/core/v1
   version: v1
 version: "3"
