+++ internal/controller/suite_test.go
@@ -1,4 +1,4 @@
-package controllers
+package controller
 
 import (
 	"path/filepath"
